quand 2 philo, le f_l est nul

+++ philo/philosophers.h
@@ -6,7 +6,7 @@
 /*   By: rhamza <rhamza@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/18 06:45:48 by rhamza            #+#    #+#             */
-/*   Updated: 2023/03/29 15:49:59 by rhamza           ###   ########.fr       */
+/*   Updated: 2023/03/29 23:40:41 by rhamza           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,7 +36,7 @@ typedef struct philo_struct
 {
     unsigned int id;
     unsigned int nb_eat;
-    unsigned int begin_activity;
+    long long int begin_activity;
     pthread_mutex_t f_r; // fourchette droite
     pthread_mutex_t *f_l; // fourchette gauche
     pthread_mutex_t *ptr_mutex_write;
